Change config-files for vins_core in vins_core/configurations folder and added timestamps calculation for video + camera handler

+++ modules/vins_core/video_source/DataPackageBase.hpp
@@ -12,7 +12,7 @@ namespace vins_core
      */
     struct DataPackageBase
     {
-        float timestamp{0};             // <- time for given image
+        double timestamp{0};             // <- time for given image
         cv::Mat img;                    // <- image from dataset
     };
